+++ ProjectSpecialK/DateTimePanel.h
@@ -15,6 +15,6 @@ private:
 public:
 	DateTimePanel();
 	void Update();
-	void Tick(double dt);
-	void Draw(double dt);
+	void Tick(float dt);
+	void Draw(float dt);
 };
