+++ ProjectSpecialK/Tickable.h
@@ -7,8 +7,8 @@ public:
 	bool dead{ false };
 
 	virtual ~Tickable() {}
-	virtual void Tick(double dt) {};
-	virtual void Draw(double dt) {};
+	virtual void Tick(float dt) {};
+	virtual void Draw(float dt) {};
 	virtual bool Character(unsigned int codepoint) { return false; }
 };
 typedef std::shared_ptr<Tickable> TickableP;
